// src/app/models/search-customer.model.ts
export class SearchCustomer {
  first_name: string;
  last_name: string;
  company: string;
  email: string;
  address_1: string;
  address_2: string;
  mobile_phone: string;
  potential_type_id: number;
  state_id: number;
  zip_code: string;
  city_id: number;
  reference_status: number;
  sub_division_name: string;
  major_intersection: string;
  created_on: string;
  fax: string;
}